refactor: custom location refactored
