+++ src/components/modals/my-setting/custom-location/LocationCascader.tsx
@@ -1,46 +1,47 @@
 import { useState, useEffect } from 'react';
 import { Button } from '@/components/ui/button';
-import { CascaderOption, DefaultCustomLocation } from '@/types/CascaderOption';
+import { CascaderOption } from '@/types/CascaderOption';
+import { CustomLocationSchema } from '@/lib/schemas';
+
+type SingleCascaderProps = {
+  prevVal: string;
+  data: CascaderOption;
+  optionList: CascaderOption[];
+  setOptionList: (state: CascaderOption[]) => void;
+  isSelected: boolean;
+  setSelected: (state: string) => void;
+};
 
 function SingleCascader({
   prevVal,
   data,
   optionList,
   setOptionList,
-  selected,
+  isSelected,
   setSelected,
-}: {
-  prevVal: string;
-  data: CascaderOption;
-  optionList: CascaderOption[];
-  setOptionList: React.Dispatch<React.SetStateAction<CascaderOption[]>>;
-  selected: boolean;
-  setSelected: React.Dispatch<React.SetStateAction<string>>;
-}) {
+}: SingleCascaderProps) {
   function clickHandler() {
-    if (selected) {
+    const buf = optionList.slice(0);
+    const idx = buf.findIndex((item) => item.value === prevVal);
+    if (isSelected) {
       setSelected('');
-      const temp = optionList.slice(0);
-      const idx = temp.findIndex((item) => item.value === prevVal);
-      temp.splice(idx + 1);
-      setOptionList(temp);
+      buf.splice(idx + 1);
+      setOptionList(buf);
       return;
     }
     setSelected(data.value);
-    const temp = optionList.slice(0);
-    const idx = temp.findIndex((item) => item.value === prevVal);
-    if (idx !== temp.length - 1) temp.splice(idx + 1);
-    temp.push(data);
-    setOptionList(temp);
+    if (idx !== buf.length - 1) buf.splice(idx + 1);
+    buf.push(data);
+    setOptionList(buf);
   }
 
   return (
     <Button
-      className={`w-24 gap-2 rounded-lg px-3 py-1 text-base md:w-32  md:text-xl ${
-        selected
-          ? 'bg-darkblue text-white hover:bg-gray-700'
-          : 'border-2 border-darkblue bg-white text-darkblue hover:bg-gray-200'
-      } border-0 border-darkblue`}
+      className={`w-24 gap-2 rounded-lg border-darkblue px-3 py-1 text-base md:w-32 md:text-xl ${
+        isSelected
+          ? 'border-0 bg-darkblue text-white hover:bg-gray-700'
+          : 'border-2 bg-white text-darkblue hover:bg-gray-200'
+      }`}
       onClick={() => clickHandler()}
     >
       {data.label}
@@ -48,27 +49,24 @@ function SingleCascader({
   );
 }
 
-function CascaderColumn({
-  item,
-  optionList,
-  setOptionList,
-  index,
-}: {
+type CascaderColumnProps = {
   item: CascaderOption;
   optionList: CascaderOption[];
-  setOptionList: React.Dispatch<React.SetStateAction<CascaderOption[]>>;
+  setOptionList: (state: CascaderOption[]) => void;
   index: number;
-}) {
+};
+
+function CascaderColumn({ item, optionList, setOptionList, index }: CascaderColumnProps) {
   const [selected, setSelected] = useState<string>('');
   const [rounded, setRounded] = useState<string>('rounded-lg');
 
   useEffect(() => {
     if (optionList.length === 1) setRounded('rounded-lg');
     else if (index === 0) setRounded('rounded-l-lg');
-    else if (index === optionList.length - 1 || !optionList[optionList.length - 1].children) {
+    else if (index === optionList.length - 1 || !optionList[optionList.length - 1].children)
       setRounded('rounded-r-lg');
-    } else setRounded('rounded-none');
-  }, [optionList]);
+    else setRounded('rounded-none');
+  }, [optionList, setRounded, index]);
 
   return (
     <span className={`flex flex-col border ${rounded}`} key={item.value}>
@@ -81,7 +79,7 @@ function CascaderColumn({
               data={child}
               optionList={optionList}
               setOptionList={setOptionList}
-              selected={selected === child.value}
+              isSelected={selected === child.value}
               setSelected={setSelected}
             />
           ),
@@ -93,17 +91,19 @@ function CascaderColumn({
 export default function LocationCascader({
   setLocationValue,
 }: {
-  setLocationValue: React.Dispatch<React.SetStateAction<string>>;
+  setLocationValue: (state: string) => void;
 }) {
-  const data = DefaultCustomLocation;
+  const data = CustomLocationSchema;
   const [optionList, setOptionList] = useState<CascaderOption[]>([data]);
   const result = optionList.map((item) => item.label).join(' ');
+
   useEffect(() => {
     setLocationValue(result);
-  }, [result]);
+  }, [result, setLocationValue]);
+
   return (
     <div className="flex flex-col gap-4">
-      <p className="text-xl  ">{result}</p>
+      <p className="text-xl">{result}</p>
       <div className="flex flex-row">
         {optionList.map(
           (item, index) =>
