+++ src/lib/clusterUtils.tsx
@@ -10,97 +10,10 @@ import {
   ActiveClusterUser,
 } from '@/types/Cluster';
 
-const clustersSchema: Clusters = {
-  c1: {
-    r1: [1, 2, 3, 4, 5, 6, 7],
-    r2: [1, 2, 3, 4, 5, 6, 7],
-    r3: [1, 2, 3, 4, 5, 6, 7],
-    r4: [1, 2, 3, 4, 5, 6, 7],
-    r5: [1, 2, 3, 4, 5, 6, 7],
-    r6: [1, 2, 3, 4, 5, 6, 7],
-    r7: [1, 2, 3, 4, 5, 6, 7],
-    r8: [1, 2, 3, 4, 5, 6, 7],
-    r9: [1, 2, 3, 4, 5, 6, 7],
-  },
-  c2: {
-    r1: [1, 2, 3, 4, 5, 6, 7, 8],
-    r2: [1, 2, 3, 4, 5, 6, 7, 8],
-    r3: [1, 2, 3, 4, 5, 6, 7, 8],
-    r4: [1, 2, 3, 4, 5, 6, 7, 8],
-    r5: [1, 2, 3, 4, 5, 6, 7, 8],
-    r6: [1, 2, 3, 4, 5, 6, 7, 8],
-    r7: [1, 2, 3, 4, 5, 6, 7, 8],
-    r8: [1, 2, 3, 4, 5, 6, 7, 8],
-    r9: [1, 2, 3, 4, 5, 6, 7, 8],
-    r10: [1, 2, 3, 4, 5, 6, 7, 8],
-  },
-  cx1: {
-    r1: [1, 2, 3, 4],
-    r2: [1, 2, 3, 4],
-    r3: [1, 2, 3, 4],
-    r4: [1, 2, 3, 4, 5, 6, 7, 8],
-    r5: [1, 2, 3, 4, 5, 6, 7, 8],
-  },
-  cx2: {
-    r1: [1, 2, 3, 4],
-    r2: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10],
-    r3: [1, 2, 3, 4, 5, 6, 7, 8],
-    r4: [1, 2, 3, 4, 5, 6],
-    r5: [1, 2, 3, 4, 5, 6],
-    r6: [1, 2, 3, 4, 5, 6, 7, 8],
-    r7: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10],
-    r8: [1, 2, 3, 4],
-  },
-  c3: {
-    r1: [1, 2, 3, 4, 5, 6, 7],
-    r2: [1, 2, 3, 4, 5, 6, 7],
-    r3: [1, 2, 3, 4, 5, 6, 7],
-    r4: [1, 2, 3, 4, 5, 6, 7],
-    r5: [1, 2, 3, 4, 5, 6, 7],
-    r6: [1, 2, 3, 4, 5, 6, 7],
-    r7: [1, 2, 3, 4, 5, 6, 7],
-    r8: [1, 2, 3, 4, 5, 6, 7],
-    r9: [1, 2, 3, 4, 5, 6, 7],
-  },
-  c4: {
-    r1: [1, 2, 3, 4, 5, 6, 7, 8],
-    r2: [1, 2, 3, 4, 5, 6, 7, 8],
-    r3: [1, 2, 3, 4, 5, 6, 7, 8],
-    r4: [1, 2, 3, 4, 5, 6, 7, 8],
-    r5: [1, 2, 3, 4, 5, 6, 7, 8],
-    r6: [1, 2, 3, 4, 5, 6, 7, 8],
-    r7: [1, 2, 3, 4, 5, 6, 7, 8],
-    r8: [1, 2, 3, 4, 5, 6, 7, 8],
-    r9: [1, 2, 3, 4, 5, 6, 7, 8],
-    r10: [1, 2, 3, 4, 5, 6, 7, 8],
-  },
-  c5: {
-    r1: [1, 2, 3, 4, 5, 6, 7],
-    r2: [1, 2, 3, 4, 5, 6, 7],
-    r3: [1, 2, 3, 4, 5, 6, 7],
-    r4: [1, 2, 3, 4, 5, 6, 7],
-    r5: [1, 2, 3, 4, 5, 6, 7],
-    r6: [1, 2, 3, 4, 5, 6, 7],
-    r7: [1, 2, 3, 4, 5, 6, 7],
-    r8: [1, 2, 3, 4, 5, 6, 7],
-    r9: [1, 2, 3, 4, 5, 6, 7],
-  },
-  c6: {
-    r1: [1, 2, 3, 4, 5, 6, 7, 8],
-    r2: [1, 2, 3, 4, 5, 6, 7, 8],
-    r3: [1, 2, 3, 4, 5, 6, 7, 8],
-    r4: [1, 2, 3, 4, 5, 6, 7, 8],
-    r5: [1, 2, 3, 4, 5, 6, 7, 8],
-    r6: [1, 2, 3, 4, 5, 6, 7, 8],
-    r7: [1, 2, 3, 4, 5, 6, 7, 8],
-    r8: [1, 2, 3, 4, 5, 6, 7, 8],
-    r9: [1, 2, 3, 4, 5, 6, 7, 8],
-    r10: [1, 2, 3, 4, 5, 6, 7, 8],
-  },
-};
+import { ClusterSchema } from '@/lib/schemas';
 
 // getClusters로 안전한 복사본 제공
-export const getClusters = (): Clusters => JSON.parse(JSON.stringify(clustersSchema));
+export const getClusters = (): Clusters => JSON.parse(JSON.stringify(ClusterSchema));
 
 // 특정 위치 업데이트 함수
 export const updateClusterUser = (
